% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polydar-package.R
\name{hardcase1}
\alias{hardcase1}
\alias{simple2d}
\title{polylidar hardcase1 and in-built simple2d}
\description{
Data set fixtures from polylidar.
}
\details{
The data are matrices of coordinates in x,y.
}
