% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polydar-package.R
\docType{package}
\name{polydar-package}
\alias{polydar}
\alias{polydar-package}
\title{polydar: Generate Polygons from Points}
\description{
Fast polygon extraction from point clouds using polylidar 
 <https://arxiv.org/abs/2002.01886>. The
 polylidar library is available from <https://github.com/JeremyBYU/polylidar>.
}
\author{
\strong{Maintainer}: Michael D. Sumner \email{mdsumner@gmail.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

Authors:
\itemize{
  \item Jeremy Castagno [copyright holder]
}

}
\keyword{internal}
